Mostly restore `hir` API boundary
